+++ test/integration/default/serverspec/default_spec.rb
@@ -2,4 +2,5 @@ require 'spec_helper'
 
 describe 'windows_schannel::default' do
   # TODO: Need to add my checks here. :)
+  # TODO: I'm still learning so happy to take a pull request if you could give me an example based on this cookbook
 end